requirements.txt
django~=1.11.29
celery~=4.4.7
pillow~=6.2.2
redis~=3.5.3
python-dotenv~=0.18.0